requirements.txt
flask
flask-cors
pillow
pandas
numpy
scikit-learn
